requirements.txt
coloredlogs==10.0
elasticsearch==7.1.0
elasticsearch-dsl==7.1.0
humanfriendly==4.18
PyMySQL==0.9.3
python-dateutil==2.8.1
PyYAML==5.1.2
six==1.13.0
urllib3==1.25.7
